verifier/csp_inconsistency.html

<!doctype html>
<meta charset=utf-8>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src="/resources/channel.sub.js"></script>
<script src=/websockets/constants.sub.js></script>
<script src=/portals/resources/stash-utils.sub.js></script>
<script src="/common/utils.js"></script>

<script>

const delay = (time) =>
      new Promise((resolve) => setTimeout(resolve, time))

setup({explicit_timeout: true});

const timeout_fn = async(time, label) => {
  await delay(time) 
  assert_true(false, label)
}

// function run_test(test_fn, timeout, label) {
//   return Promise.race(
//     [ test_fn(), 
//       timeout_fn(timeout, label)
//     ]);
// }


function run_test(timeout, label, ...fns) {
  return Promise.race(
    [ Promise.all(fns.map((x) => x())), 
      timeout_fn(timeout, label)
    ]);
}


promise_test(async t => {
  // window.setTimeout(() => { error(new AssertionError("timeout")) }, 100);

  let uuid = token();

  window.open(`server.py?corr=1&res=48&uuid=${uuid}&` + "stash=" + btoa(STASH_RESPONDER), "_blank");
  
  await run_test(1000, "Timed out", () => 
    StashUtils.takeValue(uuid).then(t.step_func_done(value => {
        assert_equals(value, "PWNED", "CSP INCONSISTENCY 1");
    })), async () => 
        assert_equals("PWNED", "PWNED", "CSP INCONSISTENCY 2")
    )

  await Promise.race(
    [ StashUtils.takeValue(uuid).then(t.step_func_done(value => {
        assert_equals(value, "PWNED", "CSP INCONSISTENCY");
      })), 
      fn(1000, "Timed out")
    ]);

}, "CSP INCONSISTENCY ATTACK TRACE");

</script>
